refactor ExecutionUtils
