+++ ui/src/components/executions/ChangeExecutionStatus.vue
@@ -76,7 +76,7 @@
     import action from "../../models/action";
     import {State} from "@kestra-io/ui-libs"
     import Status from "../../components/Status.vue";
-    import ExecutionUtils from "../../utils/executionUtils";
+    import * as ExecutionUtils from "../../utils/executionUtils";
     import {useAuthStore} from "override/stores/auth"
 
     export default {
@@ -107,7 +107,7 @@
                     })
                     .then(response => {
                         if (response.data.id === this.execution.id) {
-                            return ExecutionUtils.waitForState(this.$http, this.$store, response.data);
+                            return ExecutionUtils.waitForState(this.$http, response.data);
                         } else {
                             return response.data;
                         }
