+++ ui/src/components/executions/Resume.vue
@@ -33,7 +33,7 @@
     import action from "../../models/action";
     import {State} from "@kestra-io/ui-libs"
     import FlowUtils from "../../utils/flowUtils";
-    import ExecutionUtils from "../../utils/executionUtils";
+    import * as ExecutionUtils from "../../utils/executionUtils";
     import InputsForm from "../../components/inputs/InputsForm.vue";
     import {inputsToFormData} from "../../utils/submitTask";
     import {mapStores} from "pinia";
